Arrumando a barra de progresso e sist notificacao

+++ app/templatetags/filters.py
@@ -6,6 +6,6 @@ register = template.Library()
 @register.filter
 def divide(value, arg):
     try:
-        return int((int(value) / int(arg)) * 100)
+        return int(float(float(value)/float(arg))* 100)
     except (ValueError, ZeroDivisionError):
         return None
